requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
idna==3.10
mysqlclient==2.2.7
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
qrcode==8.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
